requirements.txt
Pillow==8.1.2
pkg-resources==0.0.0
pypng==0.0.20
PyQRCode==1.2.1
pyzbar==0.1.8
six==1.15.0
